osu-winello: added xdg-desktop-portal to deps.
Many people reported their osu! not opening in-game links due to xdg-desktop-portal missing, as Proton claims itself in the terminal output. 
Even though it should be a pretty common package, here we are 🍷

Also mentioned to install Steam on Ubuntu, Debian etc. as those distros seem to *not* pull libGL when installing drivers sometimes. Thought it was worth mentioning! 👍🏻