osu-wine: remove hide wine ver.
After latest stable update and discussing it with ppy as well, I'm removing Hide Wine as it's not supposed to be supported and it will be the cause for crashes from now on. 

Every feature it brought will soon be fixed by stable's team, hopefully the game will now run again fine!