bump proton-osu-9-10! (read for important info!)
New Proton release fixing osu! crashing when Hide Wine version is disabled!

There's been quite a fuss ever since the last stable update: it broke both
hide-wine and non-hide-wine osu!, with the second being quite catchy to fix xd

With this release, you'll notice fonts will look different from Windows' ones and
a few features not working (like boss key): osu! does that when it detects it's running
on Wine for some reason (https://github.com/NelloKudo/osu-winello/issues/115). We've already addressed it to ppy, hopefully it gets fixed soon.

In the meanwhile, for those who want the usual fonts look and features back, you can edit ~/.local/bin/osu-wine
and set WINE_BLOCK_GET_VERSION to 1, since we fixed Proton to work with it as well.
The only downside to that is that, with a future stable update, it might just break again: if that happens, please
revert it to 0 as it is the behavior osu! (and its devs) expect on Wine.

That's all! Enjoy :P