replace Error() with InstallError() in the installation script
Now that most of the post-install features are in osu-winello.sh, using
Error() for them would end up also calling the Revert() function for failed install,
hence uninstalling the whole script if any of them failed.

Solution: make an InstallError() that also calls Revert() only for the installing part
of the script, use the Error() one only for post-install functions.